+++ cinelerra/dcoffset.h
@@ -17,13 +17,13 @@ public:
 	DC_Offset();
 	~DC_Offset();
 
-	int calibrate_dc_offset(long *output, RecordGUIDCOffsetText **dc_offset_text, int input_channels);
+	int calibrate_dc_offset(int *output, RecordGUIDCOffsetText **dc_offset_text, int input_channels);
 	void run();
 
-	long *output;
+	int *output;
 	RecordGUIDCOffsetText **dc_offset_text;
 	Mutex *dc_offset_lock;
-	long dc_offset[MAXCHANNELS], dc_offset_total[MAXCHANNELS], dc_offset_count;
+	int dc_offset[MAXCHANNELS], dc_offset_total[MAXCHANNELS], dc_offset_count;
 	int input_channels;
 	int getting_dc_offset;
 	BC_ProgressBox *progress;
