+++ cinelerra/vtrack.h
@@ -27,22 +27,22 @@ public:
 	int load_defaults(Defaults *defaults);
 	void set_default_title();
 	PluginSet* new_plugins();
-	int channel_is_playable(long position, int direction, int *do_channel);
+	int channel_is_playable(int64_t position, int direction, int *do_channel);
 	int save_header(FileXML *file);
 	int save_derived(FileXML *file);
-	int load_header(FileXML *file, unsigned long load_flags);
-	int load_derived(FileXML *file, unsigned long load_flags);
+	int load_header(FileXML *file, uint32_t load_flags);
+	int load_derived(FileXML *file, uint32_t load_flags);
 	int copy_settings(Track *track);
 	void synchronize_params(Track *track);
-	long to_units(double position, int round);
+	int64_t to_units(double position, int round);
 	double to_doubleunits(double position);
-	double from_units(long position);
+	double from_units(int64_t position);
 
-	void calculate_input_transfer(Asset *asset, long position, int direction, 
+	void calculate_input_transfer(Asset *asset, int64_t position, int direction, 
 		float &in_x, float &in_y, float &in_w, float &in_h,
 		float &out_x, float &out_y, float &out_w, float &out_h);
 
-	void calculate_output_transfer(int channel, long position, int direction, 
+	void calculate_output_transfer(int channel, int64_t position, int direction, 
 		float &in_x, float &in_y, float &in_w, float &in_h,
 		float &out_x, float &out_y, float &out_w, float &out_h);
 
@@ -72,23 +72,23 @@ public:
 		float &out_y2, 
 		int frame_w, 
 		int frame_h, 
-		long real_position, 
+		int64_t real_position, 
 		int direction);
 // Give whether compressed data can be copied directly from the track to the output file
-	int direct_copy_possible(long current_frame, int direction);
+	int direct_copy_possible(int64_t current_frame, int direction);
 
 
 // ===================================== editing
 
-	int copy_derived(long start, long end, FileXML *xml);
-	int paste_derived(long start, long end, long total_length, FileXML *xml, int &current_channel);
+	int copy_derived(int64_t start, int64_t end, FileXML *xml);
+	int paste_derived(int64_t start, int64_t end, int64_t total_length, FileXML *xml, int &current_channel);
 // use samples for paste_output
-	int paste_output(long startproject, long endproject, long startsource, long endsource, int layer, Asset *asset);
-	int clear_derived(long start, long end);
-	int copy_automation_derived(AutoConf *auto_conf, long start, long end, FileXML *xml);
-	int paste_automation_derived(long start, long end, long total_length, FileXML *xml, int shift_autos, int &current_pan);
-	int clear_automation_derived(AutoConf *auto_conf, long start, long end, int shift_autos = 1);
-	int modify_handles(long oldposition, long newposition, int currentend);
+	int paste_output(int64_t startproject, int64_t endproject, int64_t startsource, int64_t endsource, int layer, Asset *asset);
+	int clear_derived(int64_t start, int64_t end);
+	int copy_automation_derived(AutoConf *auto_conf, int64_t start, int64_t end, FileXML *xml);
+	int paste_automation_derived(int64_t start, int64_t end, int64_t total_length, FileXML *xml, int shift_autos, int &current_pan);
+	int clear_automation_derived(AutoConf *auto_conf, int64_t start, int64_t end, int shift_autos = 1);
+	int modify_handles(int64_t oldposition, int64_t newposition, int currentend);
 	int draw_autos_derived(float view_start, float zoom_units, AutoConf *auto_conf);
 	int draw_floating_autos_derived(float view_start, float zoom_units, AutoConf *auto_conf, int flash);
 	int select_auto_derived(float zoom_units, float view_start, AutoConf *auto_conf, int cursor_x, int cursor_y);
@@ -99,7 +99,7 @@ public:
 // ===================================== for handles, titles, etc
 
 // rounds up to integer frames for editing
-	int identical(long sample1, long sample2);
+	int identical(int64_t sample1, int64_t sample2);
 // no rounding for drawing
 	int get_dimensions(double &view_start, 
 		double &view_units,
